API

SimulationLogs.SimulationLog — Type

SimulationLog(; values::Dict{Symbol,Vector}=Dict{Symbol,Vector}(), is_active::Bool=false)

A log of saved variables from a simulation. Variables can be accessed with dot notation or the getproperty function.

Base.values — Method

values(log::SimulationLog)

Get the value dictionary of a SimulationLog.

SimulationLogs.activate! — Function

activate!(log::SimulationLog=GLOBAL_LOG[])

Activate a SimulationLog. If no log is given, activate the GLOBAL_LOG.

SimulationLogs.deactivate! — Function

deactivate!(log::SimulationLog=GLOBAL_LOG[])

Deactivate a SimulationLog. If no log is given, deactivate the GLOBAL_LOG.

SimulationLogs.get_log — Method

get_log(sol)
get_log(sol, t)

Get the variables logged by @log from an ODESolution.

SimulationLogs.is_active — Method

is_active(log::SimulationLog)

See if simulation is currently active (able to be logged to).

SimulationLogs.reset! — Function

reset!(log::SimulationLog=GLOBAL_LOG[])

Reset the value dictionary of a SimulationLog. If no log is given, the GLOBAL_LOG will be reset

SimulationLogs.scope — Function

scope(sol, varnames)
scope!(sol, varnames)

Plot the variables in varnames from an ODESolution sol. If varnames is a vector, the variables will be plotted against time on the same axis. If varnames is a tuple, the variables will be plotted against each other.

SimulationLogs.scope! — Function

scope(sol, varnames)
scope!(sol, varnames)

Plot the variables in varnames from an ODESolution sol. If varnames is a vector, the variables will be plotted against time on the same axis. If varnames is a tuple, the variables will be plotted against each other.

SimulationLogs.@log — Macro

@log var_name = val
@log var_name expr

Log var_name to the GLOBAL_LOG. If @log is placed on an assignment for var_name, the variable will also be created in the local scope. If @log is placed with a variable name before an expression, the expression will run in the local scope without creating the variable in that scope.

Example

function lorenz!(du, u, p, t)
    @log a = u[2]-u[1]
    @log b u[3]+a
    du[1] = p[1]*a
    du[2] = u[1]*(p[2]-u[3]) - u[2]
    du[3] = u[1]*u[2] - p[3]*u[3]
end

In this example, a is evaluated into the scope of the lorenz! function and is able to be used within that scope. b, however is not evaluated into the lorenz! scope, so no b variable is created in that scope. Both a and b will be logged.